+++ frontend/src/pages/AskWrite.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import "../css/Ask.css";
+import "../style/Ask.css";
 import axios from "axios";
 
 export default function AskWrite() {
